moved states array to constants file and removed unused component
