+++ client/src/components/Card.js
@@ -8,7 +8,7 @@ import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
-import { aqiToCig, hazardLevels, monthToMonth } from '../constants';
+import { aqiToCig, hazardLevels, monthToMonth } from '../constants/constants';
 
 const useStyles = makeStyles({
   root: {
