+++ client/src/components/OneCity.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { aqiToCig, hazardLevels } from '../constants';
-import { solutionsObj } from "../constants"
+import { aqiToCig, hazardLevels } from '../constants/constants';
+import { solutionsObj } from "../constants/constants"
 
 
   const OneCity = ({ location }) => {
